// src/app/models/EnchantableType.ts
export enum EnchantableType {
  Sword = 'Sword',
  Axe = 'Axe',
  Pickaxe = 'Pickaxe',
  Shovel = 'Shovel',
  Helmet = 'Helmet',
  Chestplate = 'Chestplate',
  Leggings = 'Leggings',
  Boots = 'Boots',
  Bow = 'Bow',
  Crossbow = 'Crossbow',
  Trident = 'Trident',
  Hoe = 'Hoe',
  Fishing_Rod = 'Fishing Rod',
  Shears = 'Shears',
  Shield = 'Shield',
  Elytra = 'Elytra',
  Flint_and_Steel = 'Flint and Steel',
  Carrot_on_a_Stick = 'Carrot on a Stick',
  Warped_Fungus_on_a_Stick = 'Warped Fungus on a Stick',
  Turtle_Shell = 'Turtle Shell',
  Book = 'Book'
}